refactor(formatter): share shell scanning helpers
